Heap of work

+++ Content.Server/Procedural/Job/DungeonJob.PostGenBiome.cs
@@ -9,7 +9,7 @@ using Robust.Shared.Map;
 using Robust.Shared.Map.Components;
 using Robust.Shared.Utility;
 
-namespace Content.Server.Procedural;
+namespace Content.Server.Procedural.Job;
 
 public sealed partial class DungeonJob
 {
@@ -17,36 +17,36 @@ public sealed partial class DungeonJob
      * Handles PostGen code for marker layers + biomes.
      */
 
-    private async Task PostGen(BiomePostGen postGen, Dungeon dungeon, EntityUid gridUid, MapGridComponent grid, Random random)
+    private async Task PostGen(BiomePostGen postGen, Dungeon dungeon, Random random)
     {
-        if (_entManager.TryGetComponent(gridUid, out BiomeComponent? biomeComp))
+        if (_entManager.TryGetComponent(_gridUid, out BiomeComponent? biomeComp))
             return;
 
-        biomeComp = _entManager.AddComponent<BiomeComponent>(gridUid);
+        biomeComp = _entManager.AddComponent<BiomeComponent>(_gridUid);
         var biomeSystem = _entManager.System<BiomeSystem>();
-        biomeSystem.SetTemplate(gridUid, biomeComp, _prototype.Index(postGen.BiomeTemplate));
+        biomeSystem.SetTemplate(_gridUid, biomeComp, _prototype.Index(postGen.BiomeTemplate));
         var seed = random.Next();
         var xformQuery = _entManager.GetEntityQuery<TransformComponent>();
 
         foreach (var node in dungeon.RoomTiles)
         {
             // Need to set per-tile to override data.
-            if (biomeSystem.TryGetTile(node, biomeComp.Layers, seed, grid, out var tile))
+            if (biomeSystem.TryGetTile(node, biomeComp.Layers, seed, _grid, out var tile))
             {
-                _maps.SetTile(gridUid, grid, node, tile.Value);
+                _maps.SetTile(_gridUid, _grid, node, tile.Value);
             }
 
-            if (biomeSystem.TryGetDecals(node, biomeComp.Layers, seed, grid, out var decals))
+            if (biomeSystem.TryGetDecals(node, biomeComp.Layers, seed, _grid, out var decals))
             {
                 foreach (var decal in decals)
                 {
-                    _decals.TryAddDecal(decal.ID, new EntityCoordinates(gridUid, decal.Position), out _);
+                    _decals.TryAddDecal(decal.ID, new EntityCoordinates(_gridUid, decal.Position), out _);
                 }
             }
 
-            if (biomeSystem.TryGetEntity(node, biomeComp, grid, out var entityProto))
+            if (biomeSystem.TryGetEntity(node, biomeComp, _grid, out var entityProto))
             {
-                var ent = _entManager.SpawnEntity(entityProto, new EntityCoordinates(gridUid, node + grid.TileSizeHalfVector));
+                var ent = _entManager.SpawnEntity(entityProto, new EntityCoordinates(_gridUid, node + _grid.TileSizeHalfVector));
                 var xform = xformQuery.Get(ent);
 
                 if (!xform.Comp.Anchored)
@@ -59,15 +59,16 @@ public sealed partial class DungeonJob
             }
 
             await SuspendDungeon();
-            ValidateResume();
+            if (!ValidateResume())
+                return;
         }
 
         biomeComp.Enabled = false;
     }
 
-    private async Task PostGen(BiomeMarkerLayerPostGen postGen, Dungeon dungeon, EntityUid gridUid, MapGridComponent grid, Random random)
+    private async Task PostGen(BiomeMarkerLayerPostGen postGen, Dungeon dungeon, HashSet<Vector2i> reservedTiles, Random random)
     {
-        if (!_entManager.TryGetComponent(gridUid, out BiomeComponent? biomeComp))
+        if (!_entManager.TryGetComponent(_gridUid, out BiomeComponent? biomeComp))
             return;
 
         var biomeSystem = _entManager.System<BiomeSystem>();
@@ -95,24 +96,37 @@ public sealed partial class DungeonJob
             }
 
             await SuspendDungeon();
-            ValidateResume();
+            if (!ValidateResume())
+                return;
 
-            biomeSystem.GetMarkerNodes(gridUid, biomeComp, grid, markerTemplate, true, bounds, count,
+            biomeSystem.GetMarkerNodes(_gridUid, biomeComp, _grid, markerTemplate, true, bounds, count,
                 random, out var spawnSet, out var existing, false);
 
             await SuspendDungeon();
-            ValidateResume();
+            if (!ValidateResume())
+                return;
+
+            var checkTile = reservedTiles.Count > 0;
 
             foreach (var ent in existing)
             {
+                if (checkTile && reservedTiles.Contains(_maps.LocalToTile(_gridUid, _grid, _xformQuery.GetComponent(ent).Coordinates)))
+                {
+                    continue;
+                }
+
                 _entManager.DeleteEntity(ent);
             }
 
             await SuspendDungeon();
-            ValidateResume();
+            if (!ValidateResume())
+                return;
 
             foreach (var (node, mask) in spawnSet)
             {
+                if (reservedTiles.Contains(node))
+                    continue;
+
                 string? proto;
 
                 if (mask != null && markerTemplate.EntityMask.TryGetValue(mask, out var maskedProto))
@@ -124,14 +138,15 @@ public sealed partial class DungeonJob
                     proto = markerTemplate.Prototype;
                 }
 
-                var ent = _entManager.SpawnAtPosition(proto, new EntityCoordinates(gridUid, node + grid.TileSizeHalfVector));
+                var ent = _entManager.SpawnAtPosition(proto, new EntityCoordinates(_gridUid, node + _grid.TileSizeHalfVector));
                 var xform = xformQuery.Get(ent);
 
                 if (!xform.Comp.Anchored)
                     _transform.AnchorEntity(ent, xform);
 
                 await SuspendDungeon();
-                ValidateResume();
+                if (!ValidateResume())
+                    return;
             }
         }
     }
